+++ src/input/input.c
@@ -26,11 +26,12 @@
 #include "../engine/engine.h"
 #include "../logger/logger.h"
 #include "../logger/logger_messages_internal.h"
-
-#include "../h/time.h"
+#include "../math/vector.h"
 
 #include "input.h"
 #include "input_internal.h"
+
+#include "../h/time.h"
 
 #include <stdio.h>