+++ src/time.c
@@ -22,8 +22,8 @@
 
 #include "common/config.h"
 #include "common/limits.h"
+#include "math/math.h"
 
-#include "h/math.h"
 #include "h/time.h"
 
 #include <time.h>